Change meaning of string 'option_waiting_room'
This patch changed the meaning of the string 'option_waiting_room' to simply say 'Waiting room'.

This was missed during the implementation of commit 5b6f2166245bc1092e8347bb4643191448683e44, but should be fixed now as, on mod_form.php, the form element label and the checkbox label of the waiting room widget are saying the same.